+++ packages/react-layout/src/website/demos/item-wizard/demo.jsx
@@ -78,7 +78,8 @@ const Demo = () => {
         targetRef: refs.button,
         open: () => isSectionInvalid,
         component: FormErrors,
-        props: { 
+        props: {
+          fields: getSectionFieldsIds(sections[step]),
           onClickField: scrollToField,
         },
       },