+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Changes that have landed in master but are not yet released
 
 * Add - sections by grid css config. [Issue 62](https://github.com/yahoo/jafar/issues/62)
 
+### Documentation
+
+* Fix - demos broken links. [Issue 59](https://github.com/yahoo/jafar/issues/59)
+
 ## v1.0.6 (July 5, 2020)
 
 ### Form